Enable solution crawler in diagnostic tests
